requirements.txt
dotenv==0.9.9
duckdb==1.4.3
google-cloud-storage==3.7.0
pytest==9.0.2
google-cloud-bigquery==3.39.0